Renamed UGameBaseFrameworkSettings to UGameBaseFrameworkDeveloperSettings Created UGameBaseFrameworkGameSettings
