+++ Source/GameBaseFramework/Private/GameBaseFrameworkGameSettings.cpp
@@ -1,0 +1,6 @@
+#include "GameBaseFrameworkGameSettings.h"
+
+FName UGameBaseFrameworkGameSettings::GetCategoryName() const
+{
+    return FApp::GetProjectName();
+}
