+++ Source/GameBaseFramework/Private/GameBaseFrameworkModule.cpp
@@ -1,4 +1,4 @@
-#include "GameBaseFrameworkSettings.h"
+#include "GameBaseFrameworkDeveloperSettings.h"
 #include "IGameBaseFrameworkModule.h"
 
 IMPLEMENT_MODULE( IGameBaseFrameworkModule, GameBaseFramework )
