+++ Source/GameBaseFramework/Private/GameBaseFrameworkPCH.h
@@ -1,5 +1,5 @@
 #pragma once
 
 #include "Engine/GBFGameInstance.h"
-#include "GameBaseFrameworkSettings.h"
+#include "GameBaseFrameworkDeveloperSettings.h"
 #include "GameFramework/GBFSaveGame.h"
